update structure and dir
