+++ src/entities/attribute/index.ts
@@ -1,5 +1,4 @@
-export type * from './model/types/AttributeRequest';
-export type * from './model/types/AttributeResponse';
-export type * from './model/types/AttributeValueRequest';
-export type * from './model/types/AttributeValueResponse';
-export { attributesStore } from './model/attributeStore';
+export type * from '@/entities/attribute/model/types/attributeValueRequest.type';
+export type * from '@/entities/attribute/model/types/attributeValueResponse.type';
+export type * from '@/entities/attribute/model/types/attributeRequest.type';
+export type * from '@/entities/attribute/model/types/attributeResponse.type';
