+++ src/entities/attribute/model/attributeStore.ts
@@ -3,7 +3,7 @@ import type { DocumentTypeResponse, DocumentTypeToAttributeResponse } from '@/en
 import { getAllAttributeDocs } from '@/entities/attribute/api/api-attribute';
 import { makeAutoObservable, onBecomeObserved, runInAction } from 'mobx';
 
-import type { AttributeResponse } from './types/AttributeResponse';
+import type { AttributeResponse } from '../index';
 
 export interface CombinedAttribute extends AttributeResponse, DocumentTypeToAttributeResponse {}
