--- src/entities/attribute/model/types/AttributeResponse.ts
@@ -1,5 +1,0 @@
-export interface AttributeResponse {
-	id: number;
-	name: string;
-	type: string;
-}
