+++ src/entities/attribute/model/types/attributeRequest.type.ts
@@ -1,4 +1,4 @@
 export interface AttributeRequest {
-	name: string;
-	type: string;
+  name: string;
+  type: string;
 }
