+++ src/entities/attribute/model/types/attributeResponse.type.ts
@@ -1,0 +1,5 @@
+export interface AttributeResponse {
+  id: number;
+  name: string;
+  type: string;
+}
