+++ src/entities/attribute/model/types/attributeValueRequest.type.ts
@@ -1,0 +1,4 @@
+export interface AttributeValueRequest {
+  attributeId: number;
+  value: string;
+}
