+++ src/entities/attribute/model/types/attributeValueResponse.type.ts
@@ -1,0 +1,6 @@
+export interface AttributeValueResponse {
+  id: number;
+  name: string;
+  type: string;
+  value: string;
+}
