+++ src/entities/documents/index.ts
@@ -1,12 +1,11 @@
-export type * from './model/types/UpdateDocumentRequest';
-export type * from './model/types/DocumentFacadeResponse';
-export type * from './model/types/DocumentPermissionResponse';
-export type * from './model/types/DocumentResponse';
-export type * from './model/types/DocumentVersionResponse';
-export type * from './model/types/DocumentTypeRequest';
-export type * from './model/types/DocumentTypeToAttributeRequest';
-export type * from './model/types/DocumentTypeResponse';
-export type * from './model/types/DocumentTypeToAttributeResponse';
-export type * from './model/types/CreateDocumentRequest';
-export type * from './model/types/PatchDocumentRequest';
-export { documentTypesStore } from './model/documentTypesStore';
+export type * from '@/entities/documents/model/types/updateDocumentRequest.type';
+export type * from '@/entities/documents/model/types/documentFacadeResponse.type';
+export type * from '@/entities/documents/model/types/documentPermissionResponse.type';
+export type * from '@/entities/documents/model/types/documentResponse.type';
+export type * from '@/entities/documents/model/types/documentVersionResponse.type';
+export type * from '@/entities/documents/model/types/documentTypeRequest.type';
+export type * from '@/entities/documents/model/types/documentTypeToAttributeRequest.type';
+export type * from '@/entities/documents/model/types/documentTypeResponse.type';
+export type * from '@/entities/documents/model/types/documentTypeToAttributeResponse.type';
+export type * from '@/entities/documents/model/types/createDocumentRequest.type';
+export type * from '@/entities/documents/model/types/patchDocumentRequest.type';
