--- src/entities/documents/model/types/DocumentFacadeResponse.ts
@@ -1,8 +1,0 @@
-import type { DocumentResponse } from '@/entities/documents/model/types/DocumentResponse';
-import type { DocumentVersionResponse } from '@/entities/documents/model/types/DocumentVersionResponse';
-import type { SignatureResponse } from '@/entities/signature/model/types/SignatureResponse';
-export interface DocumentFacadeResponse {
-  document: DocumentResponse;
-  latest_version: DocumentVersionResponse;
-  signature: SignatureResponse;
-}
