--- src/entities/documents/model/types/DocumentResponse.ts
@@ -1,9 +1,0 @@
-import type { UserToDocumentResponse } from '@/entities/user/model/types/UserToDocumentResponse.ts';
-export interface DocumentResponse {
-	id: number;
-	document_type_id: number;
-	name: string;
-	document_versions_ids: number[];
-	user_permissions: UserToDocumentResponse[];
-	status: string;
-}
