--- src/entities/documents/model/types/DocumentTypeRequest.ts
@@ -1,5 +1,0 @@
-import type { DocumentTypeToAttributeRequest } from '@/entities/documents/model/types/DocumentTypeToAttributeRequest';
-export interface DocumentTypeRequest {
-  name: string;
-  attributes: DocumentTypeToAttributeRequest[];
-}
