--- src/entities/documents/model/types/DocumentTypeResponse.ts
@@ -1,6 +1,0 @@
-import type { DocumentTypeToAttributeResponse } from '@/entities/documents/model/types/DocumentTypeToAttributeResponse';
-export interface DocumentTypeResponse {
-  attributes: DocumentTypeToAttributeResponse[];
-  id: number;
-  name: string;
-}
