+++ src/entities/documents/model/types/documentPermissionResponse.type.ts
@@ -1,4 +1,4 @@
 export interface DocumentPermissionResponse {
-	id: number;
-	name: string;
+  id: number;
+  name: string;
 }
