+++ src/entities/documents/model/types/documentTypeRequest.type.ts
@@ -1,0 +1,5 @@
+import type { DocumentTypeToAttributeRequest } from '@/entities/documents/model/types/documentTypeToAttributeRequest.type.ts';
+export interface DocumentTypeRequest {
+  name: string;
+  attributes: DocumentTypeToAttributeRequest[];
+}
