+++ src/entities/documents/model/types/documentTypeToAttributeRequest.type.ts
@@ -1,4 +1,4 @@
 export interface DocumentTypeToAttributeRequest {
-	attribute_id: number;
-	is_optional: boolean;
+  attribute_id: number;
+  is_optional: boolean;
 }
