+++ src/entities/documents/model/types/documentTypeToAttributeResponse.type.ts
@@ -1,4 +1,4 @@
 export interface DocumentTypeToAttributeResponse {
-	attribute_id: number;
-	is_optional: boolean;
+  attribute_id: number;
+  is_optional: boolean;
 }
