+++ src/entities/documents/model/types/patchDocumentRequest.type.ts
@@ -1,4 +1,4 @@
-import type { AttributeValueRequest } from '@/entities/attribute/model/types/AttributeValueRequest';
+import type { AttributeValueRequest } from '@/entities/attribute/model/types/attributeValueRequest.type.ts';
 export interface PatchDocumentRequest {
   document_type_id: number;
   name: string;
