+++ src/entities/signature/index.ts
@@ -1,2 +1,2 @@
-export type * from './model/types/SignatureCreateRequest';
-export type * from './model/types/SignatureResponse';
+export type * from '@/entities/signature/model/types/signatureResponse.type';
+export type * from '@/entities/signature/model/types/signatureCreateRequest.type';
