+++ src/entities/signature/model/types/signatureResponse.type.ts
