--- src/entities/user/model/types/RegisterRequest.ts
@@ -1,5 +1,0 @@
-export interface RegisterRequest {
-	display_name: string;
-	email: string;
-	password: string;
-}
