--- src/entities/user/model/types/UserToDocumentResponse.ts
@@ -1,5 +1,0 @@
-import type { DocumentPermissionResponse } from '@/entities/documents/model/types/DocumentPermissionResponse';
-export interface UserToDocumentResponse {
-  email: string;
-  document_permissions: DocumentPermissionResponse[];
-}
