--- src/entities/user/model/types/UserUpdateRequest.ts
@@ -1,4 +1,0 @@
-export interface UserUpdateRequest {
-	display_name: string;
-	password: string;
-}
