+++ src/entities/user/model/types/authenticationRequest.type.ts
@@ -1,4 +1,4 @@
 export interface AuthenticationRequest {
-	email: string;
-	password: string;
+  email: string;
+  password: string;
 }
