+++ src/entities/user/model/types/authenticationResponse.type.ts
@@ -1,3 +1,3 @@
 export interface AuthenticationResponse {
-	token: string;
+  token: string;
 }
