+++ src/entities/user/model/types/registerRequest.type.ts
@@ -1,0 +1,5 @@
+export interface RegisterRequest {
+  display_name: string;
+  email: string;
+  password: string;
+}
