+++ src/entities/user/model/types/userToDocumentResponse.type.ts
@@ -1,0 +1,5 @@
+import type { DocumentPermissionResponse } from '@/entities/documents/model/types/documentPermissionResponse.type.ts';
+export interface UserToDocumentResponse {
+  email: string;
+  document_permissions: DocumentPermissionResponse[];
+}
