+++ src/entities/user/model/types/userUpdateRequest.type.ts
@@ -1,0 +1,4 @@
+export interface UserUpdateRequest {
+  display_name: string;
+  password: string;
+}
