--- src/entities/vote/model/types/VoteRequest.ts
@@ -1,4 +1,0 @@
-export interface VoteRequest {
-	documentId: number;
-	status: string;
-}
