--- src/entities/vote/model/types/VoteResponse.ts
@@ -1,5 +1,0 @@
-import type { VoteUserResponse } from '@/entities/vote/model/types/VoteUserResponse';
-export interface VoteResponse {
-  status: string;
-  applicationUser: VoteUserResponse;
-}
