+++ src/entities/vote/model/types/voteRequest.type.ts
@@ -1,0 +1,4 @@
+export interface VoteRequest {
+  documentId: number;
+  status: string;
+}
