+++ src/entities/vote/model/types/voteUserResponse.type.ts
@@ -1,0 +1,4 @@
+export interface VoteUserResponse {
+  email: string;
+  displayName: string;
+}
