+++ src/entities/voting/index.ts
@@ -1,2 +1,2 @@
-export type * from './model/types/VotingProcessRequest';
-export type * from './model/types/VotingProcessResponse';
+export type * from '@/entities/voting/model/types/votingProcessRequest.type';
+export type * from '@/entities/voting/model/types/votingProcessResponse.type';
