+++ src/entities/voting/model/types/votingProcessRequest.type.ts
@@ -1,0 +1,7 @@
+export interface VotingProcessRequest {
+  name: string;
+  threshold: number;
+  deadline: { [key: string]: unknown };
+  documentId: number;
+  emails: string[];
+}
