+++ src/entities/voting/model/types/votingProcessResponse.type.ts
@@ -1,4 +1,4 @@
-import type { VoteResponse } from '@/entities/vote/model/types/VoteResponse';
+import type { VoteResponse } from '@/entities/vote/model/types/voteResponse.type.ts';
 export interface VotingProcessResponse {
   id: number;
   name: string;
